requirements.txt
astroid==2.4.2
behave==1.2.6
click==7.1.2
colorama==0.4.4
extras==1.0.0
fixtures==3.0.0
Flask==1.1.2
Flask-Testing==0.8.0
fuzzywuzzy==0.18.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
linecache2==1.0.0
MarkupSafe==1.1.1
mccabe==0.6.1
mock==4.0.2
parse==1.18.0
parse-type==0.5.2
pbr==5.5.1
PyHamcrest==2.0.2
pylint==2.6.0
python-mimeparse==1.6.0
python-subunit==1.4.0
selenium==3.141.0
six==1.15.0
sure==1.4.11
testtools==2.4.0
toml==0.10.1
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1



